separate components into multiple files to match styleguide
