+++ frontend/src/components/pages/GlobalCodeReviewDashboard/CompetenceReviewQueueEntry.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import BaseReviewQueueEntry from "./BaseQueryEntry";
+
+export default function CompetenceReviewQueueEntry({ project }) {
+  return (
+    <BaseReviewQueueEntry project={project} showAllocatedReviewers={true} />
+  );
+}
