+++ frontend/src/components/pages/GlobalCodeReviewDashboard/PullRequestReviewQueueEntry.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import BaseReviewQueueEntry from "./BaseQueryEntry";
+
+export default function PullRequestReviewQueueEntry({ project }) {
+  return <BaseReviewQueueEntry project={project} />;
+}
